+++ src/commands/env/create/compute.ts
@@ -37,25 +37,13 @@ export default class EnvCreateCompute extends Command {
     alias: Flags.string({
       char: 'a',
       description: messages.getMessage('flags.alias.summary'),
-      exclusive: ['setalias'],
-    }),
-
-    setalias: Flags.string({
-      char: 'a',
-      description: messages.getMessage('flags.alias.summary'),
-      exclusive: ['alias'],
-      hidden: true,
     }),
   };
 
   async run() {
     const { flags } = await this.parse(EnvCreateCompute);
 
-    const alias = flags.alias ?? flags.setalias;
-
-    if (flags.setalias) {
-      this.warn(messages.getMessage('flags.setalias.deprecation'));
-    }
+    const alias = flags.alias;
 
     // if `--connected-org` is null here, fetchOrg will pull the default org from the surrounding environment
     const org = await fetchOrg(flags['connected-org']);